// src/types.ts
export type CreateOptions = {
  apis: API[];
  permissions: Permission[];
  locales: Locale[];
  createBackground: boolean;
  createContentScripts: boolean;
  createOptions: boolean;
  createPopup: boolean;
  directory: string;
};

export type API =
  | 'accessibilityFeatures'
  | 'action'
  | 'alarms'
  | 'bookmarks'
  | 'browserAction'
  | 'browsingData'
  | 'commands'
  | 'contentSettings'
  | 'contextMenus'
  | 'cookies'
  | 'debugger'
  | 'declarativeContent'
  | 'declarativeNetRequest'
  | 'desktopCapture'
  | 'devtools'
  | 'documentScan'
  | 'downloads'
  | 'enterprise'
  | 'extension'
  | 'fileBrowserHandler'
  | 'fileSystemProvider'
  | 'fontSettings'
  | 'gcm'
  | 'history'
  | 'i18n'
  | 'identity'
  | 'idle'
  | 'input'
  | 'loginState'
  | 'management'
  | 'notifications'
  | 'offscreen'
  | 'omnibox'
  | 'pageAction'
  | 'pageCapture'
  | 'permissions'
  | 'platformKeys'
  | 'power'
  | 'printerProvider'
  | 'printing'
  | 'printingMetrics'
  | 'privacy'
  | 'proxy'
  | 'runtime'
  | 'scripting'
  | 'search'
  | 'sessions'
  | 'sidePanel'
  | 'storage'
  | 'system'
  | 'tabCapture'
  | 'tabGroups'
  | 'tabs'
  | 'topSites'
  | 'tts'
  | 'ttsEngine'
  | 'vpnProvider'
  | 'wallpaper'
  | 'webNavigation'
  | 'webRequest'
  | 'windows';

export type Permission =
  | 'activeTab'
  | 'alarms'
  | 'background'
  | 'bookmarks'
  | 'browsingData'
  | 'certificateProvider'
  | 'clipboardRead'
  | 'clipboardWrite'
  | 'contentSettings'
  | 'contextMenus'
  | 'cookies'
  | 'debugger'
  | 'declarativeContent'
  | 'declarativeNetRequest'
  | 'declarativeNetRequestWithHostAccess'
  | 'declarativeNetRequestFeedback'
  | 'desktopCapture'
  | 'documentScan'
  | 'downloads'
  | 'downloads.open'
  | 'downloads.ui'
  | 'enterprise.deviceAttributes'
  | 'enterprise.hardwarePlatform'
  | 'enterprise.networkingAttributes'
  | 'enterprise.platformKeys'
  | 'experimental'
  | 'fileBrowserHandler'
  | 'fileSystemProvider'
  | 'fontSettings'
  | 'gcm'
  | 'geolocation'
  | 'history'
  | 'identity'
  | 'idle'
  | 'loginState'
  | 'management'
  | 'nativeMessaging'
  | 'notifications'
  | 'offscreen'
  | 'pageCapture'
  | 'platformKeys'
  | 'power'
  | 'printerProvider'
  | 'printing'
  | 'printingMetrics'
  | 'privacy'
  | 'processes'
  | 'proxy'
  | 'scripting'
  | 'search'
  | 'sessions'
  | 'sidePanel'
  | 'storage'
  | 'system.cpu'
  | 'system.display'
  | 'system.memory'
  | 'system.storage'
  | 'tabCapture'
  | 'tabGroups'
  | 'tabs'
  | 'topSites'
  | 'tts'
  | 'ttsEngine'
  | 'unlimitedStorage'
  | 'vpnProvider'
  | 'wallpaper'
  | 'webAuthenticationProxy'
  | 'webNavigation'
  | 'webRequest'
  | 'webRequestBlocking';

export type Locale =
  | 'ar'
  | 'am'
  | 'bg'
  | 'bn'
  | 'ca'
  | 'cs'
  | 'da'
  | 'de'
  | 'el'
  | 'en'
  | 'en_AU'
  | 'en_GB'
  | 'en_US'
  | 'es'
  | 'es_419'
  | 'et'
  | 'fa'
  | 'fi'
  | 'fil'
  | 'fr'
  | 'gu'
  | 'he'
  | 'hi'
  | 'hr'
  | 'hu'
  | 'id'
  | 'it'
  | 'ja'
  | 'kn'
  | 'ko'
  | 'lt'
  | 'lv'
  | 'ml'
  | 'mr'
  | 'ms'
  | 'nl'
  | 'no'
  | 'pl'
  | 'pt_BR'
  | 'pt_PT'
  | 'ro'
  | 'ru'
  | 'sk'
  | 'sl'
  | 'sr'
  | 'sv'
  | 'sw'
  | 'ta'
  | 'te'
  | 'th'
  | 'tr'
  | 'uk'
  | 'vi'
  | 'zh_CN'
  | 'zh_TW';